+++ lib/comspec/aggregate.ex
@@ -1,0 +1,11 @@
+defmodule Comspec.Aggregate do
+  @moduledoc """
+  A struct to handle an Event
+  """
+
+  use TypedStruct
+
+  typedstruct do
+    field(:name, list())
+  end
+end