+++ priv/templates/comgen.add.config/comspecs/experimental.exs
@@ -1,0 +1,21 @@
+use Mix.Config
+
+config :comspecs, Experimental,
+  spec_name: ExAccounts,
+  spec_shortdoc: "Experimental comspec",
+  spec_doc: nil,
+  spec_run_before: nil,
+  spec_run_after: nil,
+
+  aggregates: nil,
+  commands: nil,
+  command_handlers: nil,
+  command_routers: nil,
+  command_validators: nil,
+  events: nil,
+  event_handlers: nil,
+  event_projectors: nil,
+  process_managers: nil,
+  read_schemas: nil,
+  read_queries: nil
+