+++ priv/templates/comgen.build/event.ex
@@ -1,0 +1,6 @@
+# <%= event.templates.lib.dest %>
+defmodule <%= event.module_name %> do
+  @derive Json.Encoder
+
+  defstruct [<%= event.string_fields %>] 
+end