+++ priv/templates/comgen.build/event_test.ex
@@ -1,0 +1,5 @@
+# <%= event.templates.test.dest%>
+defmodule <%= event.module_name %>Test do
+  use ExUnit.Case
+
+end