+++ test/comgen_test.exs
@@ -1,8 +1,0 @@
-defmodule PhoenixCommandedTest do
-  use ExUnit.Case
-  doctest PhoenixCommanded
-
-  test "greets the world" do
-    assert PhoenixCommanded.hello() == :world
-  end
-end